Split native code into two crates
