+++ native/Cargo.toml
@@ -1,27 +1,5 @@
-[package]
-name = "karafriends"
-version = "0.1.0"
-authors = ["Connor Worley <connorbworley@gmail.com>"]
-edition = "2018"
-exclude = ["index.node"]
-
-[lib]
-name = "karafriends"
-crate-type = ["cdylib"]
-
-[dependencies]
-asio-sys = { git = "https://github.com/RustAudio/cpal", rev = "d4ba28df733fbb5e5a4d78f7a8fd789206a44cbe", optional = true }
-cpal = { git = "https://github.com/RustAudio/cpal", rev = "d4ba28df733fbb5e5a4d78f7a8fd789206a44cbe" }
-lazy_static = "1.4.0"
-neon = { version = "1.0.0-alpha.4", default-features = false, features = ["napi-8"] }
-realfft = "3.2.0"
-ringbuf = "0.3.3"
-rubato = "0.12.0"
-
-[target.'cfg(windows)'.dependencies]
-win32console = "0.1.5"
-
-[features]
-asio = ["asio-sys", "cpal/asio"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[workspace]
+members = [
+    "karafriends-lib",
+    "karafriends",
+]
